server: add proxy server for the GlobalMgr
Until we got all of the GlobalMgr packets figured out, use a proxy
server between the real GlobalMgr and anything that connects to it.
The proxy tries to parse as many packets as possible and print them
in meaningful format, but passes all data 1:1 between the upstream
and downstream connection.