packet: Refactor packet definitions
Payload is now a trait implemented by all packet structures.
All those structures can be sent directly to the PacketStream
without wrapping them in the enum.

The enum is still present as `Packet`. It's used as return
type for PacketStream recv(). It also implements Payload,
so can be sent back into the PacketStream.

There is no more dependency on `genmatch`. The packet enum
is already defined with a proc_macro #[packet_proc], so that
macro is now also generating all the impls with autogenerated
match statements.