Add base LoginMgr
A client is able to connect and send the username and password.
The LoginMgr currently prints both and lets the client disconnect
due to timeout.

The PacketStream had to be refactored from AsyncBufRead to
a custom buffered reader that provides a &mut [u8] to the inner
data, and can be later decrypted in place.